Initial attempt at a cmake build
This is based roughly on vba-m's build file and what the autotools
chain did. This (for now) completely disregards win32 building
(The visual studio project is far superior to the sdl and gtk
builds anyway). I also changed the standard extramodule include strategy
from relative to the source file's directory to relative to the
src directory. This is more robust when restructuring and simplifies
the include path specifiers