Allow LUA to link to an external lua 5.1 version
I had to change a lot in lua-engine.cpp which was
(ab)using the internal structure of lua_State.
Also there is a bitrotted embedden version of
parts of what is essentialy the bitop module.
I would really like to remove that and include the
actual module instead (or even better as an external
dependency)