Remove a pad that should never have been added
It's hard, but true! I doubt it would have had
the desired effect and always writing a u32
should be better suited for the next couple
of years (unless someone wanted to artificially
start a run before the epoch). Maybe some form
of check can be done to read 64bit timestamps,
but it has to be better than this.